Bug fix with texture drawing, but still dont know why m_height and m_width of Texture struct is not getting set properly

+++ Game/structs.h
@@ -5,8 +5,8 @@
 typedef struct Texture
 {
 	SDL_Texture* m_Texture;
-	int m_height;
-	int m_width;
+	float m_height;
+	float m_width;
 	char m_path[128];
 } Texture;
